+++ src/fr/maxlego08/menu/loader/actions/LuckPermissionSetLoader.java
@@ -1,0 +1,22 @@
+package fr.maxlego08.menu.loader.actions;
+
+import fr.maxlego08.menu.api.loader.ActionLoader;
+import fr.maxlego08.menu.api.requirement.Action;
+import fr.maxlego08.menu.api.utils.TypedMapAccessor;
+import fr.maxlego08.menu.requirement.actions.LuckPermissionSet;
+
+import java.io.File;
+
+public class LuckPermissionSetLoader implements ActionLoader {
+
+    @Override
+    public String getKey() {
+        return "permission-set,permission set,set permission,set-permission";
+    }
+
+    @Override
+    public Action load(String path, TypedMapAccessor accessor, File file) {
+        String permission = accessor.getString("permission");
+        return new LuckPermissionSet(permission);
+    }
+}